+++ libc/setjmp.h
@@ -1,0 +1,43 @@
+/*
+ *
+ * Copyright (c) 2016 Raphine Project
+ *
+ * This program is free software; you can redistribute it and/or
+ * modify it under the terms of the GNU General Public License
+ * as published by the Free Software Foundation; either version 2
+ * of the License, or (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program; if not, write to the Free Software
+ * Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
+ *
+ * Author: Liva
+ * 
+ */
+
+#ifndef __RAPH_LIBC_SETJMP_H__
+#define __RAPH_LIBC_SETJMP_H__
+
+#include <stddef.h>
+
+#ifdef __cplusplus
+extern "C" {
+#endif /* __cplusplus */
+
+  typedef struct {
+    size_t _buf[8];
+  } jmp_buf;
+  
+  int setjmp(jmp_buf env);
+  int longjmp(jmp_buf env, int val);
+
+#ifdef __cplusplus
+}
+#endif /* __cplusplus */
+
+#endif // __RAPH_LIBC_SETJMP_H__
