+++ rlib/raph.h
@@ -106,6 +106,27 @@ extern "C" {
     return data;
   }
 
+#ifdef __KERNEL__
+  static inline bool disable_interrupt() {
+    uint64_t if_flag;
+    asm volatile("pushfq; popq %0; andq $0x200, %0;":"=r"(if_flag));
+    bool did_stop_interrupt = (if_flag != 0);
+    asm volatile("cli;");
+    return did_stop_interrupt;
+  }
+  static inline void enable_interrupt(bool flag) {
+    if (flag) {
+      asm volatile("sti");
+    }
+  }
+#else
+  static inline bool disable_interrupt() {
+    return false;
+  }
+  static inline void enable_interrupt(bool flag) {
+  }
+#endif // __KERNEL__
+
 #ifdef __cplusplus
 }
 #endif
